Updated RNN to new Dataset API and changed it's estimator to use a bigger CNN and allow feeding custom LSTM states. Updated IOU algorithm for RNN_Estimator to be much cleaner and faster.

+++ supervised_vertices/rnn.py
@@ -1,3 +1,6 @@
+import sys
+
+sys.path.append('.')
 from supervised_vertices.RNN_Estimator import *
 import tensorflow as tf
 import os
@@ -11,46 +14,40 @@ def train(sess, model, training_set, validation_set, num_optimization_steps, log
     Args:
         sess: A Session.
         model: A Model.
-        optimizer: An Optimizer.
-        generator: A generator that yields `(inputs, targets)` tuples, with
             `inputs` and `targets` both having shape `[dynamic_duration, 1]`.
         num_optimization_steps: An integer.
         logdir: A string. The log directory.
     """
-
-    if os.path.exists(logdir):
-        shutil.rmtree(logdir)
 
     summary_writer = tf.summary.FileWriter(logdir=logdir, graph=sess.graph)
 
     sess.run(tf.global_variables_initializer())
     for step in range(num_optimization_steps):
         print('\rStep %d.' % (step + 1), end='')
-        durations, inputs, targets = training_set.get_batch_for_rnn(batch_size=50, max_timesteps=5)
+        durations, inputs, targets, _ = training_set.get_sample_for_rnn(max_timesteps=5)
         training_summaries, training_image_summaries, _ = sess.run(
             [model.training_summaries, model.training_image_summaries, model.train_op],
-            {model.sequence_length: durations, model.inputs: inputs,
+            {model.inputs: inputs,
              model.targets: targets})
         summary_writer.add_summary(training_summaries, global_step=step)
         summary_writer.add_summary(training_image_summaries, global_step=step)
 
         # Validate
         if step % 100 == 0:
-            durations, inputs, targets = validation_set.get_batch_for_rnn(batch_size=50, max_timesteps=5)
+            durations, inputs, targets, _ = validation_set.get_sample_for_rnn(max_timesteps=5)
             validation_summaries, validation_image_summaries = sess.run(
                 [model.validation_summaries, model.validation_image_summaries],
-                {model.sequence_length: durations, model.inputs: inputs,
-                 model.targets: targets})
+                {model.inputs: inputs, model.targets: targets})
             summary_writer.add_summary(validation_summaries, global_step=step)
             summary_writer.add_summary(validation_image_summaries, global_step=step)
 
-            # IOU
-            failed_shapes, iou = evaluate_iou(sess, model, training_set)
-            iou_summaries = sess.run(model._training_iou_summaries, {model.failures: failed_shapes, model.iou: iou})
-            summary_writer.add_summary(iou_summaries, global_step=step)
-            failed_shapes, iou = evaluate_iou(sess, model, validation_set)
-            iou_summaries = sess.run(model._validation_iou_summaries, {model.failures: failed_shapes, model.iou: iou})
-            summary_writer.add_summary(iou_summaries, global_step=step)
+        # IOU
+        if step % 1000 == 0:
+            avg_failed_shapes, avg_iou = evaluate_iou(sess, model, validation_set, batch_size=100)
+            iou_summaries = tf.Summary()
+            iou_summaries.value.add(tag='valid/failed_shapes', simple_value=avg_failed_shapes)
+            iou_summaries.value.add(tag='valid/iou', simple_value=avg_iou)
+            summary_writer.add_summary(iou_summaries, step)
 
         # Save
         if step % 1000 == 0:
@@ -58,17 +55,19 @@ def train(sess, model, training_set, validation_set, num_optimization_steps, log
 
 
 if __name__ == '__main__':
+    logdir = 'rnn'
+    do_train = False
+    restart_training = False
+
     training_set, validation_set = get_train_and_valid_datasets('dataset_polygons.npy')
-    model = RNN_Estimator(max_timesteps=5, init_scale=0.1)
+    model = RNN_Estimator()
     with tf.Session() as sess:
-        logdir = 'rnn_4'
         saver = tf.train.Saver(keep_checkpoint_every_n_hours=1, max_to_keep=2)
 
-        do_train = True
-        if do_train:
-            if os.path.exists(logdir):
-                shutil.rmtree(logdir)
+        if restart_training and os.path.exists(logdir):
+            shutil.rmtree(logdir)
             os.makedirs(logdir)
+        if do_train:
             train(sess, model, training_set, validation_set, num_optimization_steps=100000, logdir=logdir)
-
-        saver.restore(sess, save_path=logdir + '/model.ckpt')
+        elif tf.train.latest_checkpoint(logdir):
+            saver.restore(sess, save_path=logdir + '/model.ckpt')
